+++ src/com/intellij/idea/plugin/hybris/impex/completion/provider/ImpexMacrosCompletionProvider.kt
@@ -20,8 +20,7 @@ package com.intellij.idea.plugin.hybris.impex.completion.provider
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
-import com.intellij.codeInsight.lookup.LookupElementBuilder
-import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.idea.plugin.hybris.impex.codeInsight.lookup.ImpExLookupElementFactory
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroDeclaration
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.psi.util.PsiTreeUtil
@@ -38,12 +37,8 @@ class ImpexMacrosCompletionProvider : CompletionProvider<CompletionParameters>()
 
         PsiTreeUtil.findChildrenOfType(originalFile, ImpexMacroDeclaration::class.java)
             .map { it.firstChild }
-            .map { it.text }
-            .map {
-                LookupElementBuilder.create(it)
-                    .withIcon(HybrisIcons.MACROS)
-            }
-            .forEach { result.addElement(it) }
+            .map { ImpExLookupElementFactory.buildMacro(it.text) }
+            .let { result.addAllElements(it) }
     }
 
     companion object {